requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0

langgraph==0.0.62
langchain-core==0.1.0
langchain-groq==0.0.1

tavily-python==0.7.12
trafilatura==2.0.0
pypdf==6.1.0
requests==2.31.0
